FEAT: move user.model.ts and add PrivateUser interface
- User/user.model.ts -> DTO/user.model.ts
