+++ backend/src/slack/slack.service.ts
@@ -1,7 +1,7 @@
 import { WebClient } from '@slack/web-api';
 import { ResultSetHeader } from 'mysql2';
 import { executeQuery } from '../mysql';
-import * as models from '../users/users.model';
+import * as models from '../DTO/users.model';
 import UsersService from '../users/users.service';
 import { logger } from '../utils/logger';
