+++ backend/src/users/users.service.ts
@@ -1,6 +1,6 @@
 import { Like, Not } from 'typeorm';
 import * as errorCode from '../utils/error/errorCode';
-import * as models from './users.model';
+import * as models from '../DTO/users.model';
 import * as types from './users.type';
 import UsersRepository from './users.repository';
